Statistics 2.0
Contains the mostly finished statistics module. On startup the topic to
consume from is specified and the properties for creating the kafka
consumer and producer are specified in the resources folder. From this a
ConsumerHandler and  PublishHandler are started, each in their own
thread.

These Handlers share a reference to a ConcurrentHashMap which stores the
results of processing. The ConsumerHandler regularly polls the requests
topic and processsing any it finds. These results are then stored back
in the results map where the PublishHandler can find them. It regulary
checks for results to publish and if so does.

Added the docker-compose file the statisticss service. With this we can
specify how many of each statistic service we want to start using the
--scale option of docker-compose up.

Some constants and removed some logging